Raise unless iterations is given as an integer >= 1 (#2)
The documentation for `generate/3` says to increase key iterations to at
least "2^16" for passwords. When I tried passing `:math.pow(2, 16)` to
see how long it took, I was surprised to see the function hang
indefinitely.

The issue was that the generator expects to be able to subtract 1 from
the iterations count until the result equals `0`, but with a float it
would hit `0.0` and keep subtracting forever. The same problem would
happen if the iterations count were given as a negative number.